ticket:1769 - Attempting to add ADMIN permissions flag (includes ticket:1771)
By adding a flag to any permissions created via administrative permissions,
it would be possible to make that object readable. '''However''', it would
be readable by the whole group. Some logic similar to locking would be
necessary to make it readable only by the owner. This is too complicated
for an initial solution.

git-svn-id: file:///home/svn/omero/trunk@6048 05709c45-44f0-0310-885b-81a1db45b4a6