Limit `IN (?,?,?...)` size
`findRepoFiles` previously passed all the files (`basenames`)
in an entire directory into one query. This leads to strange
errors with PostgreSQL of the form:

```
IOException: Tried to send an out-of-range integer as a 2-byte value: 52671
```

Now, the list of basenames are grouped into batches, filling
up a single `Map<String, Long>` which is returned to the caller.