Switch to lookup table provider service
The previous lookup table loading implementation both leaves the task to
the individual service and has potentially severe performance
implications.  Two different loading mechanisms were used in the
thumbnail and rendering services due to the session state and service
availability.  Furthermore, the actual loading itself was deferred deep
within the renderer infrastructure.

This commit institutes a single `LutProvider` interface which can be
passed into services which desire access to the currently available
lookup tables.  A corresponding `LutProviderImpl` Spring Bean has also
been added, allowing for the implementation to be swapped in and out as
required.

Due to session containment differences between the thumbnail and
rendering services LUT loading was taken all the way.  Loading now
happens at server start time.  The only perceptible downside to this is
that adding a new LUT requires a server restart.  However, the
implementation was written in such a way that server runtime updates to
the available LUTs could be easily added if desired.